+++ examples/demo/src/games/Playground/index.js
@@ -13,7 +13,14 @@ import {
 	Tween,
 	Container,
 	Particles,
-	Blitter
+	Blitter,
+	Circle,
+	Ellipse,
+	Rectangle,
+	Triangle,
+	Polygon,
+	Star,
+	Mask
 } from 'react-phaser-bindings';
 import gemsJson from './gems.json';
 import flares from './flares.json';
@@ -120,55 +127,141 @@ class PlaygroundGame extends Component {
 	state = {
 		x: 300,
 		y: 10,
-		play: null
+		play: null,
+		flag: false
 	};
 
-	handleClick = () => {
-		console.log('?');
+	handlePlayStop = () => {
 		this.setState(({ play }) => ({
-			play: play === 'spin' ? 'finish' : 'spin'
+			play: play ? false : 'spin'
 		}));
 	};
 
+	toggleFlag = () =>
+		this.setState(({ flag }) => ({
+			flag: !flag
+		}));
+
 	render() {
-		const { x, y, play } = this.state;
+		const { x, y, play, flag } = this.state;
+
 		return (
 			<Game config={config}>
 				<Scene name="scene" assets={assets} active>
 					<Container x={0} y={0} width={100} height={100}>
-						<Tween
-							play={play}
-							animations={{
-								spin: {
-									repeat: -1,
-									// yoyo: true,
-									props: { y: -100, x: -100, opacity: 0.5 },
-									duration: 1000
-								},
-								finish: {
-									// repeat: 1,
-									// yoyo: true,
-									props: { y: 300, x: -100, opacity: 0.5 },
-									duration: 1000
-								}
-							}}
+						<Mask
+							type={flag ? 'Ellipse' : 'Rectangle'}
+							x={100}
+							y={100}
+							width={flag ? 550 : 100}
+							height={flag ? 200 : 300}
+							inverse={flag}
 						>
-							<Container x={0} y={300} width={100} height={300}>
-								<Image texture="gem" frame="prism_0002" x={0} y={30} />
-								<Image texture="gem" frame="square_0002" x={70} y={30} />
-								<Image texture="gem" frame="ruby_0000" x={2 * 70} y={30} />
-								<Image texture="gem" frame="diamond_0000" x={3 * 70} y={30} />
-								<Image texture="gem" frame="prism_0002" x={4 * 70} y={30} />
-								<Image texture="gem" frame="square_0002" x={5 * 70} y={30} />
-								<Image texture="gem" frame="ruby_0000" x={6 * 70} y={30} />
-								<Image texture="gem" frame="diamond_0000" x={7 * 70} y={30} />
-								<Image texture="gem" frame="prism_0002" x={8 * 70} y={30} />
-								<Image texture="gem" frame="square_0002" x={9 * 70} y={30} />
-								<Image texture="gem" frame="ruby_0000" x={10 * 70} y={30} />
-							</Container>
-						</Tween>
+							<Circle
+								x={100}
+								y={100}
+								radius={150}
+								fillColor="0xffff00"
+								alpha={1}
+								strokeWidth={10}
+								strokeColor="0xff0000"
+								strokeAlpha={0.7}
+								stroke={flag}
+							/>
+							<Ellipse
+								x={300}
+								y={300}
+								width={150}
+								height={100}
+								angle={30}
+								fillColor="0xff5f00"
+								strokeWidth={10}
+								strokeColor="0xff0000"
+								strokeAlpha={0.7}
+								stroke
+							/>
+							<Rectangle
+								x={560}
+								y={360}
+								width={150}
+								height={100}
+								angle={-30}
+								fillColor="0x005fff"
+								strokeWidth={3}
+								strokeColor="0x00ff00"
+								strokeAlpha={1}
+								stroke
+							/>
+							<Polygon
+								x={360}
+								y={100}
+								points={[0, 0, 100, 100, 100, 200, flag ? 400 : 300, 20]}
+								angle={0}
+								fillColor="0x005fff"
+								strokeWidth={3}
+								strokeColor="0x00ff00"
+								strokeAlpha={1}
+								stroke
+							/>
+							<Star
+								x={560}
+								y={100}
+								points={flag ? 5 : 10}
+								angle={0}
+								fillColor="0x005fff"
+								strokeWidth={3}
+								strokeColor="0x00ff00"
+								strokeAlpha={1}
+								stroke
+							/>
+							<Tween
+								play={play}
+								//onComplete={this.handlePlayStop}
+								animations={{
+									spin: {
+										complex: true,
+										queue: ['spinStart', 'spinLoop', 'spinEnd']
+									},
+									spinStart: {
+										repeat: 0,
+										props: [{ y: 250, x: 0 }, { y: 700, x: 0 }],
+										ease: 'Sine.In',
+										duration: 1000
+									},
+									spinLoop: {
+										repeat: 3,
+										props: [{ y: -100, x: 0 }, { y: 700, x: 0 }],
+										ease: 'Linear',
+										duration: 500
+									},
+									spinEnd: {
+										repeat: 0,
+										props: [{ y: -100, x: 0 }, { y: 250, x: 0 }],
+										ease: 'Sine.Out',
+										duration: 1000
+									}
+								}}
+							>
+								<Container x={0} y={250} width={100} height={300}>
+									<Image texture="gem" frame="prism_0002" x={0} y={30} />
+									<Image texture="gem" frame="square_0002" x={70} y={30} />
+									<Image texture="gem" frame="ruby_0000" x={2 * 70} y={30} />
+									<Image texture="gem" frame="diamond_0000" x={3 * 70} y={30} />
+									<Image texture="gem" frame="prism_0002" x={4 * 70} y={30} />
+									<Image texture="gem" frame="square_0002" x={5 * 70} y={30} />
+									<Image texture="gem" frame="ruby_0000" x={6 * 70} y={30} />
+									<Image texture="gem" frame="diamond_0000" x={7 * 70} y={30} />
+									<Image texture="gem" frame="prism_0002" x={8 * 70} y={30} />
+									<Image texture="gem" frame="square_0002" x={9 * 70} y={30} />
+									<Image texture="gem" frame="ruby_0000" x={10 * 70} y={30} />
+								</Container>
+							</Tween>
+						</Mask>
+						<Text interactive x={350} y={250} width={100} onClick={this.toggleFlag}>
+							PLAY / STOP
+						</Text>
 					</Container>
-					<Blitter
+					{/*<Blitter
 						x={0}
 						y={0}
 						texture="blocks"
@@ -177,14 +270,7 @@ class PlaygroundGame extends Component {
 							{
 								x: 100,
 								y: 100,
-								frame: 'block-000',
-								visible: true,
-								reset: false,
-								alpha: 1,
-								flip: {
-									x: true,
-									y: false
-								}
+								frame: 'block-000'
 							},
 							{
 								x: 200,
@@ -207,7 +293,7 @@ class PlaygroundGame extends Component {
 								frame: 'block-004'
 							}
 						]}
-					/>
+					/>*/}
 					{/* <Particles
 						texture="coins"
 						frame="Gold_21.png"